a ton of writing
Currently doesn't compile, in the middle of fixing a bug which kind of
requires taking a bunch of things apart and then putting them back
together.  Forcing myself to write the blog post explaining the code
made me realize the bug (namely, assuming that polytypes in the
context can't contain unification variables).